MBL-14755: Merged some logic in the espresso lib
--InstructureTest was merged into CanvasTest
--InstructureRunner was merged into CanvasRunner
--espresso.matchers.CustomMatchers was merged into canvas.espresso.CustomMatchers
--espresso.actions.CustomViewActions was merged into canvas.espresso.CustomActions.

refs: MBL-14755
affects: Student, Teacher
release note: Refactored some automation logic
